Move components to their respective modules.
