+++ assets/js/modules/analytics-4/components/widgets/AnalyticsPopularProductsWidget.js
