+++ assets/js/modules/analytics-4/components/widgets/AnalyticsTopCitiesWidget.test.js
@@ -1,0 +1,30 @@
+/**
+ * AnalyticsTopCitiesWidget component tests.
+ *
+ * Site Kit by Google, Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { render } from '../../../../../../tests/js/test-utils';
+import AnalyticsTopCitiesWidget from './AnalyticsTopCitiesWidget';
+
+describe( 'AnalyticsTopCitiesWidget', () => {
+	it( 'should render the widget', () => {
+		const { getByText } = render( <AnalyticsTopCitiesWidget /> );
+
+		expect(
+			getByText( 'TODO: UI for AnalyticsTopCitiesWidget' )
+		).toBeInTheDocument();
+	} );
+} );
