+++ assets/js/modules/search-console/components/widgets/SearchConsolePopularKeywordsWidget.js
