+++ assets/js/modules/search-console/components/widgets/SearchConsolePopularKeywordsWidget.test.js
@@ -1,5 +1,5 @@
 /**
- * AnalyticsTopTrafficSourceWidget component tests.
+ * SearchConsolePopularKeywordsWidget component tests.
  *
  * Site Kit by Google, Copyright 2023 Google LLC
  *
@@ -16,15 +16,15 @@
  * limitations under the License.
  */
 
-import { render } from '../../../../../tests/js/test-utils';
-import AnalyticsTopTrafficSourceWidget from './AnalyticsTopTrafficSourceWidget';
+import { render } from '../../../../../../tests/js/test-utils';
+import SearchConsolePopularKeywordsWidget from './SearchConsolePopularKeywordsWidget';
 
-describe( 'AnalyticsTopTrafficSourceWidget', () => {
+describe( 'SearchConsolePopularKeywordsWidget', () => {
 	it( 'should render the widget', () => {
-		const { getByText } = render( <AnalyticsTopTrafficSourceWidget /> );
+		const { getByText } = render( <SearchConsolePopularKeywordsWidget /> );
 
 		expect(
-			getByText( 'TODO: UI for AnalyticsTopTrafficSourceWidget' )
+			getByText( 'TODO: UI for SearchConsolePopularKeywordsWidget' )
 		).toBeInTheDocument();
 	} );
 } );
